% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\name{checklist}
\alias{checklist}
\title{Check-Lists For Either Building A Package Or Submitting To CRAN}
\usage{
checklist(type = "package")
}
\arguments{
\item{type}{Load the check-list items for package building (package"), or for
submitting to CRAN ("CRAN")?}
}
\value{
The id of the new container element, invisibly.
}
\description{
A check-list of things you should do when building a R package
from scratch, or submitting a package to CRAN.
}
\examples{
checklist("CRAN")
}
