% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_creation.R
\name{package_name_check}
\alias{package_name_check}
\title{Check If A Package Name Is Available}
\usage{
package_name_check()
}
\value{
Messages in the console.
}
\description{
Check If A Package Name Is Available
}
