% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagepal-package.R
\docType{package}
\name{packagepal-package}
\alias{packagepal}
\alias{packagepal-package}
\title{packagepal: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/packagepal}
  \item Report bugs at \url{https://github.com/lddurbin/packagepal/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
