% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{pal_checklist}
\alias{pal_checklist}
\title{A Check List For Building A Package}
\usage{
pal_checklist()
}
\value{
The id of the new container element, invisibly.
}
\description{
A check-list of things you should do when building a R package from scratch.
}
