% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_creation.R
\name{pal_create_package}
\alias{pal_create_package}
\title{Create A R Package}
\usage{
pal_create_package(package_name)
}
\arguments{
\item{package_name}{The name of the package you're going to create, passed as
a character string.}
}
\value{
Path to the newly created package, invisibly.
}
\description{
Create A R Package
}
