% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{pal_start}
\alias{pal_start}
\title{Package Prerequisites}
\usage{
pal_start()
}
\value{
Each check is called for its side-effect, returned as a message to
the console.
}
\description{
Check for package development toolchains, Git, {devtools}, and
(optionally) run a Git token check.
}
