% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagepal-package.R
\docType{package}
\name{packagepal-package}
\alias{packagepal}
\alias{packagepal-package}
\title{packagepal: A Package For Package Builders}
\description{
A suite of tools to help users on their journey towards building a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/packagepal}
  \item Report bugs at \url{https://github.com/lddurbin/packagepal/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
