% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagepal-package.R
\docType{package}
\name{packagepal-package}
\alias{packagepal}
\alias{packagepal-package}
\title{packagepal: Guidelines and Checklists for Building CRAN-Worthy Packages}
\description{
Provides essential checklists for R package developers, whether you're creating your first package or beginning a new project. This tool guides you through each step of the development process, including specific considerations for submitting your package to the Comprehensive R Archive Network (CRAN). Simplify your workflow and ensure adherence to best practices with 'packagepal'."
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/packagepal}
  \item Report bugs at \url{https://github.com/lddurbin/packagepal/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com} [copyright holder]

}
\keyword{internal}
