% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_creation.R
\name{pal_create_package}
\alias{pal_create_package}
\title{Check If A Package Name Is Available And Create It}
\usage{
pal_create_package()
}
\value{
Messages in the console.
}
\description{
Check If A Package Name Is Available And Create It
}
