% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{pal_start}
\alias{pal_start}
\title{Package Prerequisites}
\usage{
pal_start(install_missing_packages = TRUE)
}
\arguments{
\item{install_missing_packages}{Install missing packages automatically?
Defaults to {TRUE}.}
}
\value{
Each check is called for its side-effect, returned as a message to
the console.
}
\description{
Check for package development toolchains, Git, {devtools}, and
(optionally) run a Git token check.
}
